mcp_apex/requirements.txt
mcp==1.2.0
python-dotenv==1.0.1
requests==2.31.0
